updates to better organize the store files
