+++ src/stores/admin.js
@@ -3,38 +3,15 @@ import { defineStore } from 'pinia'
 export const useAdminStore = defineStore({
   id: 'admin',
   state: () => ({
-    events: [],
-    loadingEvents: false,
     loadingManagers: false,
-    managers: [],
-    spaServices: []
+    managers: []
   }),
   getters: {
-    getEvents: (state) => {
-      return state.events
-    },
     getManagers: (state) => {
       return state.managers
-    },
-    getSpaServices: (state) => {
-      return state.spaServices
     }
   },
   actions: {
-    async fetchEvents() {
-      this.loadingEvents = true
-      try {
-        await fetch('/api/events/all')
-          .then((response) => response.json())
-          .then(({ data }) => (
-            this.events = data
-          ))
-      } catch (error) {
-        console.log(error)
-      } finally {
-        this.loadingEvents = false;
-      }
-    },
     async fetchManagers() {
       this.loadingManagers = true
       try {
@@ -43,20 +20,6 @@ export const useAdminStore = defineStore({
         console.log(error)
       } finally {
         this.loadingManagers = false;
-      }
-    },
-    async fetchSpaServices() {
-      this.loadingSpaServices = true
-      try {
-        await fetch('/api/spa/services')
-          .then((response) => response.json())
-          .then(({ data }) => (
-            this.spaServices = data
-          ))
-      } catch (error) {
-        console.log(error)
-      } finally {
-        this.loadingSpaServices = false;
       }
     }
   }
